Removed extra xorshift versions, so all now use xorshift RNG.
